Started the 3 way aiming code
(Chris M) In the Character2DController script, I started writing the code that aims the bullet in one of 3 directions depending on our actions. We still have to add the code to it which instantiates bullets and gives them velocity. Also, once this is all up and running, we also have to remove the scripts that shoot multidirectionally. Also, possibly, when we're done, we might want to move the code for bullets to another script and make them public static methods.